requirements.txt
streamlit
pandas
python-dateutil
pytz
tzdata
python-docx
pypdf